+++ tests/unit/test__utils.py
@@ -18,7 +18,6 @@ from sdv._utils import (
     _get_root_tables,
     _is_datetime_type,
     _is_numerical,
-    _prepare_data_visualization,
     _validate_foreign_keys_not_null,
     check_sdv_versions_and_warn,
     check_synthesizer_version,
@@ -789,34 +788,3 @@ def test__is_numerical_string():
     # Assert
     assert str_result is False
     assert datetime_result is False
-
-
-def test__prepare_data_visualization():
-    """Test ``_prepare_data_visualization``."""
-    # Setup
-    np.random.seed(0)
-    metadata = SingleTableMetadata.load_from_dict({
-        'columns': {
-            'col1': {'sdtype': 'datetime', 'datetime_format': '%Y-%m-%d'},
-            'col2': {'sdtype': 'numerical'},
-        }
-    })
-    column_names = ['col1', 'col2']
-    sample_size = 2
-    data = pd.DataFrame({
-        'col1': ['2021-01-01', '2021-02-01', '2021-03-01'],
-        'col2': [4, 5, 6],
-    })
-
-    # Run
-    result = _prepare_data_visualization(data, metadata, column_names, sample_size)
-
-    # Assert
-    expected_result = pd.DataFrame(
-        {
-            'col1': pd.to_datetime(['2021-03-01', '2021-02-01']),
-            'col2': [6, 5],
-        },
-        index=[2, 1],
-    )
-    pd.testing.assert_frame_equal(result, expected_result)
